Move completer to deephaven_internal
Follow-up to #3114
