+++ server/src/main/java/io/deephaven/server/console/ConsoleServiceGrpcImpl.java
@@ -263,7 +263,7 @@ public class ConsoleServiceGrpcImpl extends ConsoleServiceGrpc.ConsoleServiceImp
                 safelyExecute(() -> {
                     final ScriptSession scriptSession = scriptSessionProvider.get();
                     scriptSession.evaluateScript(
-                            "from deephaven.completer import jedi_settings ; jedi_settings.set_scope(globals())");
+                            "from deephaven_internal.completer import jedi_settings ; jedi_settings.set_scope(globals())");
                     settings[0] = (PyObject) scriptSession.getVariable("jedi_settings");
                 });
                 boolean canJedi = settings[0] != null && settings[0].call("can_jedi").getBooleanValue();
